commited for testing
+ added macros for icons
+ added armed icon
* changed update function to refresh
* changed all function to refresh by default

+++ src/epd/epd4in2.cpp
@@ -251,12 +251,12 @@ void Epd::DisplayFrame(const unsigned char* frame_buffer) {
 
     if (frame_buffer != NULL) {
         SendCommand(DATA_START_TRANSMISSION_1);
-        for(int i = 0; i < width / 8 * height; i++) {
+        for(int i = 0; i < (int)(width / 8 * height); i++) {
             SendData(0xFF);      // bit set: white, bit reset: black
         }
         DelayMs(2);
         SendCommand(DATA_START_TRANSMISSION_2);
-        for(int i = 0; i < width / 8 * height; i++) {
+        for(int i = 0; i < (int)(width / 8 * height); i++) {
             SendData(pgm_read_byte(&frame_buffer[i]));
         }
         DelayMs(2);
@@ -286,7 +286,7 @@ void Epd::ClearFrame(void) {
 
 	Serial.println("enter loop");
 	DelayMs(2);
-    for(int i = 0; i < width / 8 * height; i++) {
+    for(int i = 0; i < (int)(width / 8 * height); i++) {
         SendData(0xFF);
     }
     DelayMs(2);
@@ -294,7 +294,7 @@ void Epd::ClearFrame(void) {
     DelayMs(2);
 	Serial.println("second loop");
 	DelayMs(2);
-    for(int i = 0; i < width / 8 * height; i++) {
+    for(int i = 0; i < (int)(width / 8 * height); i++) {
         SendData(0xFF);
     }
     DelayMs(2);
